Reorganize plugin's files and add empty tests
